fix: datasource rename to microsoft365_utility_microsoft_365_endpoint_reference
