+++ CHANGELOG.md
@@ -21,8 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * add label management for test failures in create-test-issues.py ([5bf33c5](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/5bf33c57c31dcf04a45a5563511eb9fa2cd98983))
 * added full unit test harness for resource microsoft365_graph_beta_identity_and_access_conditional_access_policy ([092323a](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/092323ae006a66bd7718bcbd7adf4fcd92892200))
 * added full unit test harness for resource microsoft365_graph_beta_identity_and_access_conditional_access_policy ([#1077](https://github.com/deploymenttheory/terraform-provider-microsoft365/issues/1077)) ([32039a7](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/32039a72f75d5f951a67fa45f289663c2258d6b5))
-* added new datasource utility_microsoft_365_endpoints ([75737db](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/75737dbf90f72e23844d8fc0e5d64494d9f9901d))
-* added new datasource utility_microsoft_365_endpoints ([#894](https://github.com/deploymenttheory/terraform-provider-microsoft365/issues/894)) ([e7bc9d2](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/e7bc9d267de3457cc955bfe00444b68f842b5dfd))
+* added new datasource utility_microsoft_365_endpoint_reference ([75737db](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/75737dbf90f72e23844d8fc0e5d64494d9f9901d))
+* added new datasource utility_microsoft_365_endpoint_reference ([#894](https://github.com/deploymenttheory/terraform-provider-microsoft365/issues/894)) ([e7bc9d2](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/e7bc9d267de3457cc955bfe00444b68f842b5dfd))
 * added resource android_managed_device_app_configuration_policy ([0da497e](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/0da497ecc62aefe75f7db01c527ad71355738ea6))
 * enhance acceptance tests for authentication context resource ([f166680](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/f1666809c49a7e1d4020a45026f8574af9e7d193))
 * enhance nightly tests workflow with job-level failure detection and management ([323929a](https://github.com/deploymenttheory/terraform-provider-microsoft365/commit/323929abfc323319e2f6a97709cbe7744587b21a))
