+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/responses/get_endpoints_china.json
