+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/acceptance/datasource_expressroute.tf
@@ -1,0 +1,5 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance      = "worldwide"
+  express_route = true
+}
+
