+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/acceptance/datasource_filter_exchange_optimize.tf
@@ -1,0 +1,6 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance      = "worldwide"
+  service_areas = ["Exchange"]
+  categories    = ["Optimize"]
+}
+
