+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/acceptance/datasource_required_only.tf
@@ -1,0 +1,5 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance      = "worldwide"
+  required_only = true
+}
+
