+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/acceptance/datasource_teams_media.tf
@@ -1,0 +1,6 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance      = "worldwide"
+  service_areas = ["Skype"]
+  categories    = ["Optimize"]
+}
+
