+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/unit/datasource_filter_exchange.tf
@@ -1,0 +1,5 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance      = "worldwide"
+  service_areas = ["Exchange"]
+}
+
