+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/unit/datasource_multiple_filters.tf
@@ -1,0 +1,7 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance      = "worldwide"
+  service_areas = ["Exchange", "Skype"]
+  categories    = ["Optimize", "Allow"]
+  required_only = true
+}
+
