+++ internal/services/datasources/utility/microsoft_365_endpoint_reference/tests/terraform/unit/datasource_usgov_dod.tf
@@ -1,0 +1,4 @@
+data "microsoft365_utility_microsoft_365_endpoint_reference" "test" {
+  instance = "usgov-dod"
+}
+
