--- internal/services/datasources/utility/microsoft_365_endpoints/tests/terraform/acceptance/datasource_filter_exchange_optimize.tf
@@ -1,6 +1,0 @@
-data "microsoft365_utility_microsoft_365_endpoints" "test" {
-  instance      = "worldwide"
-  service_areas = ["Exchange"]
-  categories    = ["Optimize"]
-}
-
