--- internal/services/datasources/utility/microsoft_365_endpoints/tests/terraform/acceptance/datasource_teams_media.tf
@@ -1,6 +1,0 @@
-data "microsoft365_utility_microsoft_365_endpoints" "test" {
-  instance      = "worldwide"
-  service_areas = ["Skype"]
-  categories    = ["Optimize"]
-}
-
