--- internal/services/datasources/utility/microsoft_365_endpoints/tests/terraform/unit/datasource_multiple_filters.tf
@@ -1,7 +1,0 @@
-data "microsoft365_utility_microsoft_365_endpoints" "test" {
-  instance      = "worldwide"
-  service_areas = ["Exchange", "Skype"]
-  categories    = ["Optimize", "Allow"]
-  required_only = true
-}
-
