--- internal/services/datasources/utility/microsoft_365_endpoints/tests/terraform/unit/datasource_required_only.tf
@@ -1,5 +1,0 @@
-data "microsoft365_utility_microsoft_365_endpoints" "test" {
-  instance      = "worldwide"
-  required_only = true
-}
-
