--- internal/services/datasources/utility/microsoft_365_endpoints/tests/terraform/unit/datasource_usgov_dod.tf
@@ -1,4 +1,0 @@
-data "microsoft365_utility_microsoft_365_endpoints" "test" {
-  instance = "usgov-dod"
-}
-
